Avoid downloading/extracting when not needed. Added release steps option

+++ build-scripts/Release.ps1
@@ -5,9 +5,9 @@ Param (
   [Parameter(HelpMessage = "The variant for this release (alpha, stable, rc.0, rc.1).")]
   [string]$variant = "alpha",
 
-  [Parameter(HelpMessage = "The variant for this release (alpha, stable, rc.0, rc.1).")]
-  [ValidateSet('uwp', 'wasm', 'sdk', 'docs')]
-  [string[]]$build = @('uwp', 'wasm', 'sdk', 'docs'),
+  [Parameter(HelpMessage = "If supplied, only the release steps supplied will be performed.")]
+  [ValidateSet('clean', 'uwp', 'wasm', 'sdk', 'docs', 'snapshotrepo', 'snapshotdeps', 'versionbumps',  'organize', 'publish')]
+  [string[]]$steps = @('clean', 'uwp', 'wasm', 'sdk', 'docs', 'snapshotrepo', 'snapshotdeps', 'versionbumps', 'organize', 'publish'),
 
   [Parameter(HelpMessage = "The git remote to use for snapshotting the repo, pushing tags, version updates, snapshotted dependency data, and generated changelogs.")]
   [string]$gitRemote,
@@ -38,50 +38,56 @@ Param (
 # - Require an installation of IPFS for most of it.
 
 #################
-# Cleanup / Setup
+# Cleanup
 #################
-Write-Output "Cleaning up build directory"
-mkdir -Force "$PSScriptRoot/build" | Out-Null;
-Get-ChildItem "$PSScriptRoot/build" | Remove-Item –Recurse -Force -ErrorAction SilentlyContinue
+if ($steps.Contains("clean")) {
+  Write-Output "Cleaning up build directory"
+  mkdir -Force "$PSScriptRoot/build" | Out-Null;
+  Get-ChildItem "$PSScriptRoot/build" | Remove-Item –Recurse -Force -ErrorAction SilentlyContinue
+}
 
 #################
 # Version bumps
 #################
-Write-Output "Bumping version and generating changelog for Strix Music SDK"
-$sdkTag = &"$PSScriptRoot\CreateSdkRelease.ps1" -variant $variant -dryRun | Select-Object -Last 1
-$sdkChangelogLastOutput = &"$PSScriptRoot\GenerateChangelogs.ps1" -variant $variant -target sdk -forceTag $sdkTag | Select-Object -Last 1
-$emptySdkChangelog = $sdkChangelogLastOutput.ToLower().Contains("no changes");
-
-if (!$emptySdkChangelog) {
-  # Excluding -dryRun allows creation of tags and writing to disk.
-  &"$PSScriptRoot\CreateSdkRelease.ps1" -variant $variant
-}
-
-Write-Output "Bumping version and generating changelog for Strix Music App"
-$appTag = &"$PSScriptRoot\CreateAppRelease.ps1" -variant $variant -dryRun | Select-Object -Last 1
-$appChangelogLastOutput = &"$PSScriptRoot\GenerateChangelogs.ps1" -variant $variant -target app -forceTag $appTag | Select-Object -Last 1
-$emptyAppChangelog = $appChangelogLastOutput.ToLower().Contains("no changes");
-
-if (!$emptyAppChangelog) {
-  # Excluding -dryRun allows creation of tags and writing to disk.
-  &".\CreateAppRelease.ps1" -variant $variant
+if ($steps.Contains("versionbumps")) {
+    Write-Output "Bumping version and generating changelog for Strix Music SDK"
+    $sdkTag = &"$PSScriptRoot\CreateSdkRelease.ps1" -variant $variant -dryRun | Select-Object -Last 1
+    $sdkChangelogLastOutput = &"$PSScriptRoot\GenerateChangelogs.ps1" -variant $variant -target sdk -forceTag $sdkTag | Select-Object -Last 1
+    $emptySdkChangelog = $sdkChangelogLastOutput.ToLower().Contains("no changes");
+    
+    if (!$emptySdkChangelog) {
+      # Excluding -dryRun allows creation of tags and writing to disk.
+      &"$PSScriptRoot\CreateSdkRelease.ps1" -variant $variant
+    }
+    
+    Write-Output "Bumping version and generating changelog for Strix Music App"
+    $appTag = &"$PSScriptRoot\CreateAppRelease.ps1" -variant $variant -dryRun | Select-Object -Last 1
+    $appChangelogLastOutput = &"$PSScriptRoot\GenerateChangelogs.ps1" -variant $variant -target app -forceTag $appTag | Select-Object -Last 1
+    $emptyAppChangelog = $appChangelogLastOutput.ToLower().Contains("no changes");
+    
+    if (!$emptyAppChangelog) {
+      # Excluding -dryRun allows creation of tags and writing to disk.
+    &".\CreateAppRelease.ps1" -variant $variant
+  }
 }
 
 #################
 # Snapshot dependencies
 #################
-# Download build dependencies, upload to IPFS, and update the CIDs and URLs in dependencies.json.
-Write-Output "Snapshotting dotnet binaries"
-.\SnapshotDotnetSdk.ps1
-
-Write-Output "Extracting and testing dotnet binary"
-.\dotnet.ps1 -Command "--version" -fallbackOnly
-
-Write-Output "Snapshotting nuget packages needed to build WebAssembly"
-.\SnapshotNugetPackages.ps1 -skipDownload -skipExtract
-
-Write-Output "Creating snapshot of docfx binaries"
-.\RestoreDependencies.ps1 -dependencyName "docfx"
+if ($steps.Contains("snapshotdeps")) {
+  # Download build dependencies, upload to IPFS, and update the CIDs and URLs in dependencies.json.
+  Write-Output "Snapshotting dotnet binaries"
+  .\SnapshotDotnetSdk.ps1
+  
+  Write-Output "Extracting and testing dotnet binary"
+  .\dotnet.ps1 -Command "--version" -fallbackOnly
+  
+  Write-Output "Snapshotting nuget packages needed to build WebAssembly"
+  .\SnapshotNugetPackages.ps1 -skipDownload -skipExtract
+  
+  Write-Output "Creating snapshot of docfx binaries"
+  .\RestoreDependencies.ps1 -dependencyName "docfx"
+}
 
 #################
 # Commit changes
@@ -127,35 +133,37 @@ if (($noPublish -eq $false) -and ("" -ne $gitRemote) -and (!$emptyAppChangelog -
 #################
 # Snapshot git repo
 #################
-# This should be done after everything is committed.
-Write-Output "Create snapshot of git repo"
-.\SnapshotGitRepo.ps1 -outputPath $PSScriptRoot/build/source
+if ($steps.Contains("snapshotrepo")) {
+  # This should be done after everything is committed.
+  Write-Output "Create snapshot of git repo"
+  .\SnapshotGitRepo.ps1 -outputPath $PSScriptRoot/build/source
+}
 
 #################
 # Compile
 #################
 # Build documentation website
-if ($build.Contains("docs")) {
+if ($steps.Contains("docs")) {
   Write-Output "Generating documentation"
   .\GenerateDocs.ps1 -fallbackOnly
 }
 
-if ($build.Contains("sdk")) {
+if ($steps.Contains("sdk")) {
   # Create SDK nuget package
   Write-Output "Building the Strix Music SDK in $configuration mode"
-  .\dotnet.ps1 -Command 'build "$PSScriptRoot/../src/Sdk/StrixMusic.Sdk/StrixMusic.Sdk.csproj" -c $configuration' -skipExtract -skipDownload
+  .\dotnet.ps1 -Command 'build "../src/Sdk/StrixMusic.Sdk/StrixMusic.Sdk.csproj" -c $configuration' -fallbackOnly
 
   Write-Output "Packing the Strix Music SDK in $configuration mode"
-  .\dotnet.ps1 -Command 'pack "$PSScriptRoot/../src/Sdk/StrixMusic.Sdk/StrixMusic.Sdk.csproj" -c $configuration --output build/sdk/$sdkTag' -skipExtract -skipDownload
+  .\dotnet.ps1 -Command 'pack "../src/Sdk/StrixMusic.Sdk/StrixMusic.Sdk.csproj" -c $configuration --output build/sdk/$sdkTag' -fallbackOnly
 }
 
-if ($build.Contains("wasm")) {
+if ($steps.Contains("wasm")) {
   # Build WebAssembly
   Write-Output "Building WebAssembly app in $configuration mode"
-  .\dotnet.ps1 -Command 'build $PSScriptRoot/../src/Platforms/StrixMusic.Wasm/StrixMusic.Wasm.csproj /r /p:Configuration="$configuration"' -skipDownload -skipExtract
+  .\dotnet.ps1 -Command 'build $PSScriptRoot/../src/Platforms/StrixMusic.Wasm/StrixMusic.Wasm.csproj /r /p:Configuration="$configuration"' -fallbackOnly
 }
 
-if ($build.Contains("uwp")) {
+if ($steps.Contains("uwp")) {
   # Build UWP (Requires Windows with correct tooling installed)
   Write-Output "Cleaning up existing uwp AppPackages"
   Get-ChildItem "$PSScriptRoot/../src/Platforms/StrixMusic.UWP/AppPackages/" | Remove-Item -Recurse -Force
@@ -167,28 +175,32 @@ if ($build.Contains("uwp")) {
 #################
 # Organize
 #################
-# The resulting folder can be uploaded anywhere (not just ipfs)
-Write-Output "Organizing generated release content"
-.\OrganizeReleaseContent.ps1 -wasmAppPath "$PSScriptRoot/../src/Platforms/StrixMusic.Wasm/bin/x64/$configuration/net7.0/dist/*" -uwpSideloadBuildPath "$PSScriptRoot/../src/Platforms/StrixMusic.UWP/AppPackages/*" -websitePath $PSScriptRoot/../www/* -docsPath $PSScriptRoot/../docs/wwwroot/* -sdkNupkgFolder $PSScriptRoot/build/sdk/$sdkTag/* -cleanRepoPath $PSScriptRoot/build/source -buildDependenciesPath $PSScriptRoot/build/dependencies/* -outputPath $outputPath
+if ($steps.Contains("organize")) {
+  # The resulting folder can be uploaded anywhere (not just ipfs)
+  Write-Output "Organizing generated release content"
+  .\OrganizeReleaseContent.ps1 -wasmAppPath "$PSScriptRoot/../src/Platforms/StrixMusic.Wasm/bin/x64/$configuration/net7.0/dist/*" -uwpSideloadBuildPath "$PSScriptRoot/../src/Platforms/StrixMusic.UWP/AppPackages/*" -websitePath $PSScriptRoot/../www/* -docsPath $PSScriptRoot/../docs/wwwroot/* -sdkNupkgFolder $PSScriptRoot/build/sdk/$sdkTag/* -cleanRepoPath $PSScriptRoot/build/source -buildDependenciesPath $PSScriptRoot/build/dependencies/* -outputPath $outputPath
 
-if ($pastReleaseCid.Length -gt 0 -or $pastReleaseIpns.Length -gt 0) {
-  # Grab previous versioned content such as nuget packages and app installers (requires ipfs)
-  Write-Output "Importing CIDs of previous releases from $pastReleaseIpns/versions.json"
-  .\ImportPreviousVersionedContent.ps1 -url $pastReleaseIpns -cid $pastReleaseCid -outputPath $outputPath
-}
-else {
-  Write-Warning "No past release was specified, so no past release CIDs will be added to versions.json. Specify -pastReleaseIpns <addr> or -pastReleaseCid <cid> to import past published releases.";
+  if ($pastReleaseCid.Length -gt 0 -or $pastReleaseIpns.Length -gt 0) {
+    # Grab previous versioned content such as nuget packages and app installers (requires ipfs)
+    Write-Output "Importing CIDs of previous releases from $pastReleaseIpns/versions.json"
+    .\ImportPreviousVersionedContent.ps1 -url $pastReleaseIpns -cid $pastReleaseCid -outputPath $outputPath
+  }
+  else {
+    Write-Warning "No past release was specified, so no past release CIDs will be added to versions.json. Specify -pastReleaseIpns <addr> or -pastReleaseCid <cid> to import past published releases.";
+  }
 }
 
 #################
 # Publish!
 #################
-if ($ipnsPublishKey.Length -le 0) {
-  Write-Warning "No ipns publish key provided, content will not be published. Specify the name of an IPNS key with -ipnsPublishKey KeyName to add generated content to IPFS and publish to IPNS."
-}
-elseif ($noPublish -eq $false) {
-  Write-Output "Publishing to IPFS"
-  .\PublishToIpfs.ps1 $outputPath -ipnsKey $ipnsPublishKey
+if ($steps.Contains("publish")) {
+  if ($ipnsPublishKey.Length -le 0) {
+    Write-Warning "No ipns publish key provided, content will not be published. Specify the name of an IPNS key with -ipnsPublishKey KeyName to add generated content to IPFS and publish to IPNS."
+  }
+  elseif ($noPublish -eq $false) {
+    Write-Output "Publishing to IPFS"
+    .\PublishToIpfs.ps1 $outputPath -ipnsKey $ipnsPublishKey
+  }
 }
 
 Write-Output "Release completed"
